feat: add Factory Pattern class
Added a new directory containing the primary class used by
the Factory Pattern Design. It includes a static method
getInstance() to create a single instance for the factory,
another method registerStrategy() to keep the required strategies
stored in a std::map, and one last method chooseStrategy(),
which will be designed to choose dynamically between the different
strategies.