feat: Implement methods from StrategyFactory class
The following functions have been implemented in the StrategyFactory.cpp
file:
getInstance() -> This function returns a static instance for the StrategyFactory
 class. This ensures that we will register all strategies in the same instance.
registerStrategy() -> This function register a strategy into the map.
chooseStrategy() -> This function returns the desired strategy if it is found, and NULL if the strategy is not found in the map.