feat: Add factory pattern working with blocks
The factory pattern has been refactored to work with directives and
blocks. Now the complete logic is separated in two different functions:

factoryCheckDirectiveCase: Manage the logic related to parse directives.
It takes the pointer to the directive and the iterator, and search the
requested strategy inside the factory map.

factoryCheckBlockCase: Manage the logic related to parse Blocks. It
taker only the iterator to keep track of the current block, and search
the requested strategy inside the factory block map.