feat: Add config builder interface
I have created the interface used by all the builders.
It has a setDirective() method used for the dispatch table to find the needed
directive handle.
Also a addNestedBuilder() to assure the hierarchy is the same than in
composite.
And a build method that set default values that doesn't been
initializes.