feat: Add Dispatch Table base class
I created the class to manage handles for each directive in a dynamic
way. It is a template class, which means it will have access to the
derived class that inherits from it. It uses the Curiously
Recurring Template Pattern (CRTP) to achieve it.

The registerHandler function will store each handler function
inside a map, which key values for each pointer to a handler function
will be the name of the directive.

Later, during the Builder process, the dispatchDirective function will
be finding in the map the needed handler.