refactor: Change parameters methods IConfigBuilder
I have changed some of the methods in IConfigBuilder to match the
necessities of the project. Now the build method takes an AConfigBlock
and will be the one who will call all the handles and the
addNestedBuilder function.

The addNestedBuilder function now takes also the AConfigBlock of the
current child.

Also I have added a new method called setDefaultValues, which have the
old functioning of build.