refactor: Change structures and errorPages to map
The structures t_return and t_errorPage have change to fit the
requirements of the data to be stored. I have added boolean attributes
for checking that the structs have been set.

Now errorPages is a map of status codes pointing to some url or uri. All
they have a bool to check if the equal modifier has been set.