feat: inherance with SOLID
Now class inherance is deleted and is in the proper configs files where
the inherance are managed. I extracted all structs to another new file
to make the program compileable. Iconfig now have two virtual to manage
all inherence in class, but the next step is to achieve that all
function manage with virtual function. Right now, the herenci of block
is hardcodinf.