feat: added listen and epoll to socketsManager
Added three functions to the socketsManager class so now it listens to the ports and initializes the epoll instance. The listen function is called in the main loop of the server, and the epoll instance is used to monitor the sockets for incoming connections. New runEpollLoop in the WEbserv.run. Tested OK.