feat: parser HTTP and some handle
I implemented a parser that take a string and analise it in two parts.
In the first part, it take three string: method, path and version. In
the second part it kept all header in a map container with two
std::string. In adition, I made two handle to manage Host and
contentLegth. To the next commit i wish to make the other handle in the
future.