feat: request parser added
Implemented a system to infinite bucle to manage request headers and
body with a manager request class. This class build and request class
that manage basics headers and build a std::string body. In future push,
i will change body to make it a vector of char.
The decidion of take only basics headers is to allow impleted other
headers when it is necesary.