feat: Added php cgi to execute
Now the user registration and login is done with php and sets a cookie for the user. Small change to add .php handling in the cgi part. Aslo small changes in the .html to call the .php and added a new unmatched.conf that works as a basic server test.