feat: Improved send() for bigger file sizes
Now we use the own epoll status to send the response in parts of a buffer if the file size is too big for just one send, so now the remote connection to the server works perfectly. Also some structural changes so the non blocking socket could handle this new way of sending information.